Refactor Folder Structure
handled each request in seprate file.
