+++ src/controllers/Auth/authController.js
@@ -1,0 +1,12 @@
+const handleLogin = require('./handleLogin')
+const handleLogOut = require('./handleLogOut')
+const handleSignUp = require('./handleSignUp')
+const handleRefresh = require('./handleRefresh')
+
+module.exports = {
+    handleLogin,
+    handleLogOut,
+    handleSignUp,
+    handleRefresh
+}
+
