+++ src/controllers/Bus/busController.js
@@ -1,0 +1,4 @@
+const handleListBuses = require('./handleListBuses')
+const handleGetLocations = require('./handleGetLocations')
+
+module.exports = {handleListBuses, handleGetLocations}
