+++ src/model/dbUtils.js
@@ -1,5 +1,3 @@
-
-
 const reqString = {
     type : String,
     required : true
