+++ src/routes/bookings.js
@@ -4,15 +4,14 @@ const router = express.Router();
 
 // router for /bookings
 
-
 //     handle following routes here
 
-// POST    /bookings/new 
-// GET     /bookings/list
-// GET     /bookings/:bookingId
-// POST    /bookings/mail
-// POST    /bookings/cancel/initiate
-// POST    /bookings/cancel
+// TODO :: POST    /bookings/new 
+// TODO :: GET     /bookings/list
+// TODO :: GET     /bookings/:bookingId
+// TODO :: POST    /bookings/mail
+// TODO :: POST    /bookings/cancel/initiate
+// TODO :: POST    /bookings/cancel
 
 
 router.get('/', (req, res) => res.send("Bookings Router"));
